+++ packages/components/ng-ovh-otrs/src/otrs-popup/otrs-popup.controller.js
@@ -1,9 +1,29 @@
-angular.module("ovh-angular-otrs").controller("OtrsPopupCtrl", function ($rootScope, $stateParams, $translate, $q, $transitions, OvhApiMeVipStatus, OvhApiMe, OvhApiSupport, OvhApiProductsAapi, OtrsPopupService, UNIVERSE,
-                                                                         TICKET_CATEGORIES, OTRS_POPUP_ASSISTANCE_ENUM, OTRS_POPUP_BILLING_ENUM, OTRS_POPUP_INCIDENT_ENUM, OTRS_POPUP_INTERVENTION_ENUM, OTRS_POPUP_UNIVERSES) {
+angular.module("ovh-angular-otrs").controller("OtrsPopupCtrl", function ($q, $rootScope, $scope, $stateParams, $transitions, $translate,
+                                                                         OvhApiMe, OvhApiMeVipStatus, OvhApiProductsAapi, OvhApiSupport,
+                                                                         OtrsPopupService, OtrsPopupInterventionService,
+                                                                         OTRS_POPUP_ASSISTANCE_ENUM, OTRS_POPUP_BILLING_ENUM, OTRS_POPUP_CATEGORIES, OTRS_POPUP_INCIDENT_ENUM, OTRS_POPUP_INTERVENTION_ENUM, OTRS_POPUP_SERVICES, OTRS_POPUP_UNIVERSES,
+                                                                         TICKET_CATEGORIES, UNIVERSE) {
     "use strict";
 
     var self = this;
     var OTHER_SERVICE = "other";
+
+    self.loaders = {
+        send: false,
+        services: false,
+        models: false,
+        intervention: false
+    };
+    self.currentUser = null;
+    self.isVIP = false;
+    self.services = [];
+
+    $transitions.onSuccess({}, function (transition) {
+        var toParams = transition.params();
+        if (toParams.projectId && self.services && self.services.indexOf(toParams.projectId) !== -1) {
+            self.ticket.serviceName = toParams.projectId;
+        }
+    });
 
     function initFields () {
         self.ticket = {
@@ -18,6 +38,23 @@ angular.module("ovh-angular-otrs").controller("OtrsPopupCtrl", function ($rootSc
         var standardizedUniverse = UNIVERSE === "GAMMA" ? "SUNRISE" : UNIVERSE;
 
         self.selectedUniverse = _.includes(["CLOUD", "DEDICATED"], standardizedUniverse) || !standardizedUniverse ? "CLOUD_DEDICATED" : standardizedUniverse;
+
+        self.intervention = {
+            serviceName: null,
+            request: "intervention",
+            disk: {
+                comment: null,
+                disks: [{
+                    id: 1
+                }],
+                inverse: false
+            },
+            hasMegaRaid: false,
+            slotInfo: {},
+            enums: {
+                slotID: []
+            }
+        };
     }
 
     function manageAlert (message, type) {
@@ -31,13 +68,51 @@ angular.module("ovh-angular-otrs").controller("OtrsPopupCtrl", function ($rootSc
         self.alert.type = type;
     }
 
-    $transitions.onSuccess({}, function (transition) {
-        var toParams = transition.params();
-        if (toParams.projectId && self.services && self.services.indexOf(toParams.projectId) !== -1) {
-            self.ticket.serviceName = toParams.projectId;
-        }
-    });
+    self.getServices = function () {
+        self.loaders.services = true;
 
+        // hide alert
+        manageAlert();
+        return OvhApiProductsAapi
+            .get({
+                includeInactives: true,
+                universe: self.selectedUniverse === "CLOUD_DEDICATED" ? "DEDICATED" : self.selectedUniverse
+            }).$promise.then(function (services) {
+                var translationPromises = services.results.map(function (s) {
+                    return $translate("otrs_service_category_" + s.name, null, null, s.name).then(function (value) {
+                        s.translatedName = value;
+                        return s;
+                    });
+                });
+
+                return $q.all(translationPromises).then(function (services) { // eslint-disable-line no-shadow
+                    services = services.sort(function (a, b) { return a.translatedName.localeCompare(b.translatedName); }); // eslint-disable-line no-param-reassign
+                    services.push({
+                        translatedName: $translate.instant("otrs_service_category_other"),
+                        services: [
+                            {
+                                serviceName: OTHER_SERVICE,
+                                displayName: $translate.instant("otrs_service_category_other")
+                            }
+                        ]
+                    });
+                    self.services = services;
+
+                    return services;
+                });
+            })
+            .catch(function (err) {
+                manageAlert([($translate.instant("otrs_err_get_infos"), err.data && err.data.message) || ""].join(" "), "danger");
+            })
+            .finally(function () {
+                self.loaders.services = false;
+            });
+    };
+
+    /**
+     * Send ticket.
+     * @return {Promise}
+     */
     self.sendTicket = function () {
         // hide alert
         manageAlert();
@@ -50,70 +125,147 @@ angular.module("ovh-angular-otrs").controller("OtrsPopupCtrl", function ($rootSc
                 self.ticket.category = TICKET_CATEGORIES.DEFAULT;
             }
 
-            OvhApiSupport.Lexi().create(self.ticket).$promise.then(
-                function (data) {
+            return OvhApiSupport.v6()
+                .create(self.ticket).$promise
+                .then(function (data) {
                     initFields();
                     self.otrsPopupForm.$setUntouched();
                     self.otrsPopupForm.$setPristine();
                     $rootScope.$broadcast("ticket.otrs.reload");
-                    manageAlert($translate.instant("otrs_popup_sent_success", { ticketNumber: data.ticketNumber, ticketId: data.ticketId }), "success");
-                },
-                function (err) {
+                    manageAlert($translate.instant("otrs_popup_sent_success", {
+                        ticketNumber: data.ticketNumber,
+                        ticketId: data.ticketId
+                    }), "success");
+                })
+                .catch(function (err) {
                     manageAlert([($translate.instant("otrs_popup_sent_error"), err.data && err.data.message) || ""].join(" "), "danger");
-                }
-            ).finally(function () {
-                self.loaders.send = false;
+                })
+                .finally(function () {
+                    self.loaders.send = false;
+                });
+        }
+
+        return $q.when();
+    };
+
+    /**
+     * Send disk replacement.
+     * @return {Promise}
+     */
+    self.sendDiskReplacement = function () {
+        if (!self.loaders.send) {
+            self.loaders.send = true;
+
+            return OtrsPopupInterventionService
+                .sendDiskReplacement(self.intervention.serviceName, self.intervention.disk)
+                .then(function (data) {
+                    initFields();
+                    $rootScope.$broadcast("ticket.otrs.reload");
+                    manageAlert($translate.instant("otrs_popup_sent_success", {
+                        ticketNumber: data.ticketNumber,
+                        ticketId: data.ticketId
+                    }), "success");
+                }).catch(function (err) {
+                    if (_.includes(err.message, "This feature is currently not supported in your datacenter")) {
+                        manageAlert($translate.instant("otrs_popup_sent_error_not_available"), "danger");
+                    } else if (_.includes(err.message, "Action pending : ticketId ")) {
+                        var ticketId = /\d+$/.exec(err.message);
+                        manageAlert($translate.instant("otrs_popup_sent_error_already_exists", {
+                            ticketId: ticketId
+                        }), "danger");
+                    } else {
+                        manageAlert([$translate.instant("otrs_popup_sent_error"), _.get(err, "message", "")].join(" "), "danger");
+                    }
+                }).finally(function () {
+                    self.loaders.send = false;
+                    self.refreshRequests();
+                    self.setForm("start");
+                });
+        }
+
+        return $q.when();
+    };
+
+    function getSelectedService () {
+        return _.filter(self.services, function (service) {
+            return _.find(service.services, { serviceName: self.ticket.serviceName });
+        });
+    }
+
+    function isSelectedChoiceDedicatedServer () {
+        return _.first(_.map(getSelectedService(), "name")) === "SERVER";
+    }
+
+    self.refreshRequests = function () {
+        if (isSelectedChoiceDedicatedServer() && !_.includes(self.requests, self.intervention.request)) {
+            self.requests.push(self.intervention.request);
+        } else if (!isSelectedChoiceDedicatedServer()) {
+            _.remove(self.requests, function (req) {
+                return self.intervention.request === req;
             });
+            self.ticket.category = undefined;
+            self.ticket.subcategory = null;
         }
     };
 
+    self.refreshFormDetails = function () {
+        if (self.ticket.subcategory === OTRS_POPUP_INTERVENTION_ENUM.REPLACEMENTDISK && self.formDetails === "message") {
+            self.refreshRequests();
+            self.setForm("start");
+        }
+    };
 
-    self.getServices = function () {
-
-        self.loaders.services = true;
-        self.services = [];
-
-        // hide alert
-        manageAlert();
-
-        OvhApiProductsAapi.get({
-            includeInactives: true,
-            universe: self.selectedUniverse === "CLOUD_DEDICATED" ? "DEDICATED" : self.selectedUniverse
-        }).$promise.then(function (services) {
-            var translationPromises = services.results.map(function (s) {
-                return $translate("otrs_service_category_" + s.name, null, null, s.name).then(function (value) {
-                    s.translatedName = value;
-                    return s;
-                });
+    function getServerInfo () {
+        self.loaders.intervention = true;
+        self.intervention.canHotSwap = false;
+        self.intervention.hasMegaRaid = false;
+        self.intervention.slotInfo.canUseSlotId = null;
+        self.intervention.slotInfo.slotsCount = 0;
+        return OtrsPopupInterventionService
+            .getServerInterventionInfo(self.intervention.serviceName)
+            .then(function (serverInfo) {
+                self.intervention.canHotSwap = serverInfo.canHotSwap;
+                self.intervention.hasMegaRaid = serverInfo.hasMegaRaid;
+                self.intervention.slotInfo = serverInfo.slotInfo;
+                if (self.intervention.slotInfo.canUseSlotId) {
+                    self.intervention.enums.slotID = Array.apply(null, { length: self.intervention.slotInfo.slotsCount }).map(Number.call, Number);
+                }
+            }).catch(function () {
+                manageAlert($translate.instant("otrs_intervention_disk_error"), "danger");
+            }).finally(function () {
+                self.loaders.intervention = false;
             });
+    }
 
-            return $q.all(translationPromises).then(function (services) { // eslint-disable-line no-shadow
-                services = services.sort(function (a, b) { return a.translatedName.localeCompare(b.translatedName); }); // eslint-disable-line no-param-reassign
+    self.addDisk = function () {
+        var newItemNo = self.intervention.disk.disks.length + 1;
+        self.intervention.disk.disks.push({ id: newItemNo });
+    };
 
-                services.push({
-                    translatedName: $translate.instant("otrs_service_category_other"),
-                    services: [
-                        {
-                            serviceName: OTHER_SERVICE,
-                            displayName: $translate.instant("otrs_service_category_other")
-                        }
-                    ]
-                });
-                self.services = services;
-            });
-        })
-            .catch(function (err) { manageAlert([($translate.instant("otrs_err_get_infos"), err.data && err.data.message) || ""].join(" "), "danger"); })
-            .finally(function () { self.loaders.services = false; });
+    self.removeChoice = function (item) {
+        self.intervention.disk.disks.splice(item, 1);
+    };
+
+    self.setForm = function (formDetails) {
+        self.formDetails = formDetails;
+    };
+
+    self.continueForm = function () {
+        if (self.ticket && self.ticket.category === OTRS_POPUP_CATEGORIES.INTERVENTION && self.ticket.subcategory === OTRS_POPUP_INTERVENTION_ENUM.REPLACEMENTDISK) {
+            self.intervention.serviceName = self.ticket.serviceName;
+            self.setForm(OTRS_POPUP_INTERVENTION_ENUM.REPLACEMENTDISK);
+        } else if (self.ticket && self.ticket.category === OTRS_POPUP_CATEGORIES.INTERVENTION && self.ticket.subcategory === OTRS_POPUP_INTERVENTION_ENUM.OTHER) {
+            self.ticket.category = OTRS_POPUP_CATEGORIES.INCIDENT;
+            self.ticket.subcategory = OTRS_POPUP_INCIDENT_ENUM.DOWN;
+            self.setForm("message");
+        } else {
+            self.setForm("message");
+        }
     };
 
     this.$onInit = function () {
 
         initFields();
-
-        self.loaders = {
-            send: false,
-            models: true
-        };
 
         self.alert = {
             visible: false,
@@ -121,17 +273,26 @@ angular.module("ovh-angular-otrs").controller("OtrsPopupCtrl", function ($rootSc
             message: null
         };
 
-        self.isVIP = false;
-
-        self.getServices();
+        self.servicesValues = OTRS_POPUP_SERVICES;
+        self.formDetails = "start";
+        self.interventionEnum = OTRS_POPUP_INTERVENTION_ENUM;
 
         // hide alert
         manageAlert();
 
-        $q.all([OvhApiMe.Lexi().get().$promise, OvhApiSupport.Lexi().schema().$promise]).then(function (data) {
-            self.types = data[1].models["support.TicketTypeEnum"].enum;
-            self.categories = data[1].models["support.TicketProductEnum"].enum;
-            self.requests = data[1].models["support.TicketCategoryEnum"].enum;
+        return $q.all({
+            services: self.getServices(),
+            me: OvhApiMe.v6().get().$promise,
+            meVipStatus: OvhApiMeVipStatus.v6().get().$promise,
+            supportSchema: OvhApiSupport.v6().schema().$promise
+        }).then(function (results) {
+            self.currentUser = results.me;
+
+            self.isVIP = _.values(results.meVipStatus.toJSON()).indexOf(true) !== -1;
+
+            self.types = results.supportSchema.models["support.TicketTypeEnum"].enum;
+            self.categories = results.supportSchema.models["support.TicketProductEnum"].enum;
+            self.requests = results.supportSchema.models["support.TicketCategoryEnum"].enum;
 
             self.subCategories = {
                 assistance: [
@@ -148,12 +309,11 @@ angular.module("ovh-angular-otrs").controller("OtrsPopupCtrl", function ($rootSc
                     OTRS_POPUP_INCIDENT_ENUM.DOWN
                 ],
                 intervention: [
-                    OTRS_POPUP_INTERVENTION_ENUM.REPLACEMENTDISK,
-                    OTRS_POPUP_INTERVENTION_ENUM.OTHER
+                    OTRS_POPUP_INTERVENTION_ENUM.REPLACEMENTDISK
                 ]
             };
 
-            if (data[0].ovhSubsidiary !== "FR") {
+            if (self.currentUser.ovhSubsidiary !== "FR") {
                 self.subCategories.assistance.splice(2, 0, OTRS_POPUP_ASSISTANCE_ENUM.NEW);
                 self.subCategories.assistance.splice(3, 0, OTRS_POPUP_ASSISTANCE_ENUM.OTHER);
                 self.subCategories.billing.splice(1, 0, OTRS_POPUP_ASSISTANCE_ENUM.NEW);
@@ -164,14 +324,23 @@ angular.module("ovh-angular-otrs").controller("OtrsPopupCtrl", function ($rootSc
             if (self.categories.length === 1) {
                 self.ticket.product = self.categories[0];
             }
-        }
-        )
+
+            $scope.$watch("OtrsPopupCtrl.intervention.serviceName", function (server, oldServer) {
+                if (server && server !== oldServer) {
+                    getServerInfo();
+                }
+            });
+
+            $scope.$watch("OtrsPopupCtrl.ticket.serviceName", function () {
+                self.refreshFormDetails();
+                self.refreshRequests();
+            });
+
+            $scope.$watch("OtrsPopupCtrl.ticket.subcategory", function () {
+                self.refreshFormDetails();
+            });
+        })
             .catch(function (err) { manageAlert([($translate.instant("otrs_err_get_infos"), err.data && err.data.message) || ""].join(" "), "danger"); })
             .finally(function () { self.loaders.models = false; });
-
-        OvhApiMeVipStatus.Lexi().get().$promise.then(function (vipStatus) {
-            self.isVIP = _.values(vipStatus.toJSON()).indexOf(true) !== -1;
-        });
     };
-
 });