+++ packages/components/ng-ovh-otrs/src/otrs-popup/otrs-popup.service.js
@@ -1,9 +1,9 @@
 angular.module("ovh-angular-otrs")
-    .service("OtrsPopupService", function ($rootScope, $compile) {
+    .service("OtrsPopupService", function ($rootScope) {
         "use strict";
 
         var self = this;
-        var actions = ["minimize", "maximize", "restore", "close"];
+        var actions = ["minimize", "maximize", "restore", "close", "open"];
         var isLoaded = false;
 
         angular.forEach(actions, function (action) {
@@ -13,18 +13,16 @@ angular.module("ovh-angular-otrs")
         });
 
         this.toggle = function () {
-            if ($("[otrs-popup] .draggable").hasClass("close")) {
-                self.restore();
+            if ($("[data-otrs-popup] .draggable").hasClass("close")) {
+                self.open();
             } else {
                 self.close();
             }
         };
 
         this.init = function () {
-            $compile("<div otrs-popup></div>")($rootScope, function (el) {
-                $("body").append(el);
-                isLoaded = true;
-            });
+            self.open();
+            isLoaded = true;
         };
 
         this.isLoaded = function () {