+++ packages/components/ng-ovh-otrs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [4.0.0] - 2018-03-30
+- Add missing intervention feature
+### Breaking change
+- requires ovh-api-services v3.0.0
+
 ## [3.0.0] - 2017-09-07
 - Show all services of all universes for ticket creation.
 - Add universe filter for ticket creation.
@@ -11,5 +16,3 @@
 ## [2.0.0] - 2017-08-31
 - Every service has been prefixed by OvhApi.
 - Upgrade ovh-api-services to version 2.
-
-